fix: add common dto & add date validation
