+++ src/common/common.dto.ts
@@ -1,0 +1,15 @@
+import { IsDateString, IsOptional, IsString, Matches } from 'class-validator';
+
+export class GetCountDto {
+  @IsOptional()
+  @IsString()
+  @Matches(/^\d{4}-\d{2}-\d{2}$/)
+  @IsDateString({ strict: true })
+  from?: string;
+
+  @IsOptional()
+  @IsString()
+  @Matches(/^\d{4}-\d{2}-\d{2}$/)
+  @IsDateString({ strict: true })
+  to?: string;
+}
