+++ src/user/user.dto.ts
@@ -10,7 +10,6 @@ import {
   IsNumber,
   IsOptional,
   IsString,
-  Matches,
   Validate,
 } from 'class-validator';
 import { IsPositiveInteger, IsRecordStringNumber } from '../common/validators';
@@ -197,16 +196,4 @@ export class CreateUserResponseDto {
   @IsNotEmpty()
   @IsNumber()
   userId: number;
-}
-
-export class GetCountDto {
-  @IsOptional()
-  @IsString()
-  @Matches(/^\d{4}-\d{2}-\d{2}$/)
-  from?: string;
-
-  @IsOptional()
-  @IsString()
-  @Matches(/^\d{4}-\d{2}-\d{2}$/)
-  to?: string;
 }
